Fix Product Mapper

+++ server/src/main/java/com/server/app/repository/ProductRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product,Integer> {
+public interface ProductRepository extends JpaRepository<Product, Short> {
 
     Optional<Product> findProductByProductId(Short id);
